yolo-lambda-inference/requirements.txt
flask
ultralytics==8.3.16
Pillow==11.0.0
numpy
opencv-python-headless
